move utility methods into shared utils package, extract shared merging functionality
